Move source files into core/ and audio/ folders
Organize modules into logical groups:
- src/core/ — protocol, state, websocket, time sync
- src/audio/ — decoder, scheduler

All import paths updated. No logic changes.
