+++ src/core/websocket-manager.ts
